requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
supabase==2.15.1
python-dotenv==1.1.0
google-auth==2.39.0
google-auth-oauthlib==1.2.2
google-api-python-client==2.169.0
pymupdf==1.25.5
requests==2.31.0